requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==4.3.0
black==24.1.1
boto3==1.34.0
botocore==1.34.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
constantly==23.10.4
coverage==7.4.1
cryptography==44.0.3
distlib==0.3.9
Django==5.2.1
django-crispy-forms==2.4
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.109.2
filelock==3.18.0
flake8==7.0.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
isort==5.13.2
jmespath==1.0.1
logzero==1.7.0
mangum==0.17.0
mccabe==0.7.0
mypy==1.8.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
pipenv==2025.0.3
platformdirs==4.3.8
pluggy==1.5.0
pandas
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclean==3.1.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic-settings==2.1.0
pydantic_core==2.33.2
pyflakes==3.2.0
pymongo==3.11.0
pyOpenSSL==25.1.0
pyotp==2.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.28.2
rsa==4.9.1
s3transfer==0.9.0
service-identity==24.2.0
six==1.17.0
smartapi-python==1.5.5
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.36.3
structlog==24.1.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.27.1
virtualenv==20.31.2
websocket-client==1.8.0
zope.interface==7.2
